backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
sqlalchemy>=2.0.23
sqlmodel>=0.0.14
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
bcrypt==4.0.1
google-generativeai>=0.3.1
arxiv>=2.1.0
requests>=2.31.0
aiohttp>=3.9.1
# PyMuPDF==1.23.8  # Requires Visual Studio Build Tools on Windows
pypdf2>=3.0.1
pdfplumber>=0.10.3
beautifulsoup4>=4.12.2
# lxml>=4.9.3  # Requires compilation on Windows
pydantic>=2.5.0
pydantic-settings>=2.1.0